Update Octavia keypair when public key changed
A change in the desired public key was not being reflected in the
Overcloud on stack update/upgrade. TripleO should replace the keypair
even when the desired public key changed, i.e. its fingerprint does not
match the one (possibly) already existing in the overcloud. We should
compare fingerprints and replace (delete and create, no keypair update
option) when they mismatch.

Closes-Bug: #1861031

Change-Id: I953c35c9ec24844598108bc173e84868393a98aa