Rise async value for containers starting commands
Official documentation states:

If the value of async: is not high enough, this will cause the "check
on it later" task to fail because the temporary status file that the
async_status: is looking for will not have been written or no longer
exist.

And the listed there example has async=1000.

Rize the value of a 300 used in tripleo_container_manage for creating
new containers to appear to be "high enough"...

Change-Id: I598d230bfbb5696f1c0799295d9e456eade56676
Closes-bug: #1861093
Signed-off-by: Bogdan Dobrelya <bdobreli@redhat.com>