Remove the tripleo_job_ansible_args default
This change removes the default '-v' from the environment variable when executing
local tests. This was done to enhance output for the end user by default but the
side effect was that not scenario tests were being executed. This change will ensure
any developers running the local tests are getting the same output we'd expect in
upstream zuul.

Change-Id: I76a1e44cea8140a6760e21946d62a0ecac368a24
Signed-off-by: Kevin Carter <kecarter@redhat.com>