Simplify ssh_known_hosts role
This is only required for compute nodes running nova_migration_target so
we can simplify the logic significantly.

The host.network entries are now omitted as cold/live migration only uses
either the fqdn,short hostname, or IP. This should help a little with scaling
too as ssh_host_keys can get gigantic with a large number of computes.

We can assume the remaining vars for networks and fqdn/ip are all set as host
or role group vars in the inventory.
Just in case fallback to basic entry when the host vars are missing.

This should also make it easier for operators to run the role in isolation
e.g to quickly fix up the ssh keys on any compute hosts omittied from a
scale-up.

Also fixes bug #1861296 which was caused by attempting to use set to override
a jinja2 var from an outer scope.

Change-Id: I5c91122b6cbd731d369b19b13fd011114dd48175
Depends-On: https://review.opendev.org/#/c/704919/
Closes-bug: #1861296