Octavia: do not attempt container restart on empty list
With composable roles, the node with the octavia agents may not have any
processes running when we hit step 5 external deploy tasks, resulting in
our "one line" restart failing. This patch splits the query for octavia
containers out into a separate task and we conditionally run the restart
task depending on whether there are containers found.

Change-Id: I08a06b300d43aca7db2c9e3341a07bc903163d26
Closes-Bug: #1861413