tripleo-container-manage: introduce tripleo_container_manage_valid_exit_code
If tripleo_container_manage_valid_exit_code is set to a list of codes,
for each container that will be created and expected to run something,
Ansible will check that the container exited with a valid code, given in
the parameter.

It can be useful for the Puppet containers where the entrypoint returns
the code from Puppet that has run in the container.

Change-Id: I268d7b257334d78da11c37cb0ba0783fbe2021c0