Also use enabled_networks
Since we need to generate /etc/hosts on the undercloud as well[1], we
have to use the enabled_networks in addition to role_networks, else we might
lack some entries on the overcloud nodes, namely the internalapi part.

It also makes the generated file a bit less "full of blank", using the
"-" flag in some specific places.

[1] https://review.opendev.org/705634

Change-Id: Ie942e25ed3501cd4988597ea5dd8d1a17eceec67
Related-Bug: #1861782