Use ANSIBLE_PRIVATE_KEY_FILE when adding hosts to inventory
The cli command openstack overcloud admin authorize allows the user to
specify any ssh key with --overcloud-ssh-key. That key ends up getting
set as the environment variable ANSIBLE_PRIVATE_KEY_FILE.

When cli-enable-ssh-admin.yaml runs and adds the overcloud nodes to the
inventory, it needs to use the key from ANSIBLE_PRIVATE_KEY_FILE, if one
was specified.

This patch also removes strategy:free from the play to apply the
tripleo_create_admin role as it is not compatible with any_errors_fatal,
which is desired. Using the free strategy also results in more difficult
to parse output in that the task results aren't grouped per task.

Change-Id: Ib29b2050bf02561a9cd3d96716fc6fa3c4ea3994