module tripleo_baremetal_expand_roles
This module consumes the baremetal deployment yaml and transforms it
to a list of instances and a heat environment. Subsequent modules in
the role will consume the instances and take actions based on the
instance entries and the observed ironic nodes.

The mistral action[1] has significant unit test coverage[2] which is
desirable to keep, so most of the logic has been put in a module_utils
with its own unit tests. Some special handling is required in the unit
test and doc generation to make baremetal_deploy importable from
ansible.module_utils, which is done automatically during ansible runs.

[1] https://opendev.org/openstack/tripleo-common/src/branch/master/tripleo_common/actions/baremetal_deploy.py#L353
[2] https://opendev.org/openstack/tripleo-common/src/branch/master/tripleo_common/tests/actions/test_baremetal_deploy.py#L648

Change-Id: I21cc6939db5120d4c1549b9ec66d6e0f172fd229
Story: 2007212
Task: 38457