Stop docker service when is available only.
When purging the docker service/directories, if there has
been an operating system upgrade in between (RHEL7 to
RHEL8), then the docker service will be inactive but the
folders used by Docker still need a purge. If the service
is not available and we try to stop it, then the Ansible
service task will fail.

This patch adds a guard to stop the service only when it
is available in the system.

Change-Id: Id6dc9a7bd71a7725ae1672c91bef5064a730b530