tripleo_container_manage: reduce cleanup when single config is given
If a single JSON file is given to tripleo_container_manage_config, we
don't want to remove containers from the same config in that directory,
we just want to manage that container. Which is often used to either
debug or just controlling one container at a time (FFU).

This patch will change the needs_delete() filter, and if only one config
is given, we won't cleanup other containers.

We also add the clean_orphans config param to the method, so we can
override it later.

Also if no config is given (could be due to a wrong given
tripleo_container_manage_config_patterns which would not match
anything), we don't want to remove anything for safety purpose.

Change-Id: Ic06ac0f41767ca513c612b1ebb38d2ff92500ea5