Use domain variable instead of generic resource data
When migrating from puppet to ansible, some variables were changed to be
more specific. Included in this was the variable for domain creation,
which is used when deploying keystone with domain-specific backend
template (e.g., backing keystone domains to LDAP servers):

  https://review.opendev.org/#/c/701026/19

Recently, domain information was put into
batched_tripleo_keystone_resources_domains instead of
batched_tripleo_keystone_resources_data. This commit updates the domain
plays to use the new variable. Not doing this will result in an error
like the following on the undercloud:

  fatal: [undercloud]: FAILED! => {"msg": "'batched_tripleo_keystone_resources_data' is undefined"}

Change-Id: I33530e6e3f7ac156e07545b0455c99bb7d2274ef