container_puppet_config: update config_hash for all steps
Every time the container_puppet_config module will be called, we will go
over all container startup config and update the config hash if the
config volume was changed.

It'll make sure that configs are always up to date; so we avoid useless
container restarts after an initial deployment.

Also improve the logging to show the container name when config hash
changed, instead of this whole path.

Change-Id: I67ec7ad661c0649a3448437397c0de55be984589