Improve execution and add a port check
This change introduces a fact check for neutron ports which will allow us
to pull a list of used IP addresses from our known port list which is
contains fixed_addresses. This port list will then be used to determine the
default ssh_user when first running a deployment. By pulling the neutron
facts and using the information to dictact the access user we'll be able
to support both pre-provisioned nodes and ironic provisioned nodes at the
same time within the same playbook.

The cli-enable-ssh-admin.yaml makes several API intensive calls to heat
and neutron, so to speed things up we're using ansible async. This change
moves our api intensive calls to the top of the playbooks and blocks on
their completion before they're needed. By doing this we'll improve the
overall playbook execution time.

Closes-Bug: #1863920
Change-Id: Ib79747ee7212534ab7c58d8a3e0e1d33f6069485
Depends-On: I221480f3cfc77545a8fcbef777829239c3bad0a0
Signed-off-by: Kevin Carter <kecarter@redhat.com>