tripleo_container_manage: reduce downtime for minor updates
If a container has a new config (e.g. new image), it'll now be removed
right before being re-created.

Before, we were first removing that container among potential orphan
containers, and the container would be re-created later but the downtime
in the middle can be several minutes; since we batch the container
creation.

Now, we separate the cleanup of orphan containers and the ones with new
configs.

The workflow is the following:

1) Remove orphan containers (not in the config anymore, missing Labels,
   etc).

Then for each batch of containers:

2) Remove the containers where the config_data has changed.

3) Create the containers that are in the config.

This patch should reduce the downtime of containers that are updated.

Change-Id: I821d674dead4a21b7ac30b47b31b8dd34e0ecc8b