tripleo-container-manage: remove hashed-* from default
Thanks to https://review.opendev.org/#/c/706972/ (and the module
itself), we don't have hashed files anymore. The hash is now in the
container file, and we don't have 2 files anymore; which makes things
easier to understand and debug.

This patch updates the doc and the defaults.
For backward compatibility and upgrades, if a file has hashed in its
name, it's still supported and it'll be clean-up.

Change-Id: Ie01d5509a479a726adac6490a5f32ef2df2ccde6
(cherry picked from commit 5770707777db7758aede694c4319525bf0e76727)