tripleo_container_manage: create a filter to get changed containers
Getting to know which containers changed is a challenge since we create
the containers in async with podman_container.

This patch introduces a new filter, called get_changed_containers() which
knows how to use returned data from async and figure out what containers
have podman actions (which mean podman rm and create); which have to
trigger the systemd playbook and make sure the services are triggering
changes in systemd.

Change-Id: I2ecf7d942dcca1e381d329e11939fd31299551f5