Allow to update container startup config with new config hashes
1) container_puppet_config: introduce update_config_hash_only

If update_config_hash_only is set to True, the module will do the
following:

Browse the container startup configs and for each config, check if the
config_hash has changed (e.g. Puppet has run before and generated a new
config, therefore the container will need a restart); then update the
startup configs with the new hash.

This extends container_puppet_config capabilities instead of writting a
new module for that.

2) tripleo_container_manage: add tripleo_container_manage_check_puppet_config

tripleo_container_manage_check_puppet_config is a new parameter, that is
set to False by default but if set to True, we will call the
container_puppet_config module with update_config_hash_only set to True
so we get the new config hashes in the container startup configs right
before we decide if a container needs to be restarted.

Change-Id: I16b2972cdf79cd6ac927925607197ec2a969a28b