tripleo_container_manage: disable logging for tasks using container_data
container_data contains the config data for each container, which can be
passwords or other secrets. Ansible would show it in the logs when
executing the resources, which leaks sensitive data. Only enable that
display if tripleo_container_manage_debug is set to True; like we
already do for some other tasks.

Change-Id: I2d58a3a234b94d5b5d2cc16b47c9751f78a416d9