paunch: fail early if containers weren't deployed by Paunch before
If an operator runs an update or upgrade of a cloud where containers
were deployed by Ansible and not Paunch; and accidentally forget an
environment file or a configuration somewhere (e.g. EnablePaunch); it
might result of Paunch being called.
To prevent that, Paunch will fail to run if
/var/lib/tripleo-config/.ansible-managed file is detected; let's do the
same in the Ansible module to fail early.

Change-Id: I19c8d41cad860202351035e58fe336bb8577bcad