Add a new set of tasks to build the openstack pool and key list
Using this approach we can build the list of the openstack
pools and keys according to the tht variables generated and
the list of the enabled services provided by config-download.
This patch also contains 02ea16 that fixes the way we compute
the enabled_services list on the overcloud and the condition
to check if a service is present in the list.

Closes-Bug: #1864477
Closes-Bug: #1867004
Change-Id: I80ac1fc325a2d13be1a1e454eee0389510c090b2
(cherry picked from commit 144f5aa690eb2c6ea6071dc2732c721e6bde4c77)

Get the list of enabled_services on the overcloud

This patch fixes the way we compute the enabled_services
list on the overcloud.
We don't need to use hiera anymore since the overcloud
services are already provided by config-download.
This patch also fixes the condition to check if a service
is present in the list.

Change-Id: I9952a4df122c00c0163b7e79dab8a821e3feff06
(cherry picked from commit 02ea16fbf2f631869e9366e757d80c71eb33befd)