tripleo_container_manage: add systemd testcase for service reload
Add a test that update a container managed by systemd; with a new image.
The container should be removed and re-created; and its systemd files.

We have seen race conditions in Ansible/Systemd where the unit file
could not be found. This patch will prevent that to happen thanks to a
functional test.

Change-Id: I7b7392f68d7f5b4b991c5238e7abd40d72a08b10