Ensure ruamel is included when we build test environments
This change ensures that we include ruamel.yaml when we build test
environments, which is required for docs. This module used to be
included elsewhere however, due to recent dependency changes the
package was dropped somwhere along the way.

The docs test will now run anytime change change is made to our
requirements files. This will ensure we're not running into these
types of failures in the future.

Change-Id: If3d65944a4e71fc8fba926fd0956a73429fbb172
Signed-off-by: Kevin Carter <kecarter@redhat.com>