Remove tripleo_ceph_fetch_dir role
The state information in the ceph-ansible fetch directory
is no longer necessary to replace a Ceph monitor as of
ceph-ansible 4.0.14-1. Stein, Train, Ussuri and newer
should be using this version of ceph-ansible or newer.
TripleO does not need to do extra work to maintain the
fetch directory. Remove tripleo_ceph_fetch_dir role and
modify code flow so that it's no longer used.

Also, remove unnecessary dependencies in tripleo-ceph-*
roles as described in LP #1844837.

Change-Id: Ic292dcc1bcf1e7575285ad56d56df2966e968b6d
Depends-On: I3f8c2571a6dd35029d71dd42df637a962d9b3235
Related-Bug: #1863809
Closes-Bug: #1844837