container_puppet_config: disable SElinux labelling
This is a leftover from container-puppet.py that was missed in the
migration to Ansible. If we don't do that, SElinux will report alerts
because the entrypoint (container-puppet.sh) tries to copy (and
therefore read) the content of /tmp/puppet-tmp which isn't readable per
current OpenStack SElinux podman policies.

While this is done in the policy, let's port the configuration that was
done before in container-puppet.py and this patch will be reverted once
the policy is less restritive.

Change-Id: I5baefe16f313cc17d369c9f16529516fc8d6f6e5
Related-Bug: #1864501
(cherry picked from commit 99448e20feabf141a49d6953d980c2de6a7c4b07)