Make run-local.yml use ansible_connection local, not ssh
If the intent of run-local.yml is to use a local ansible
connection and not an SSH ansible connection, as implied
by the tests/hosts.ini, then it shouldn't set a fact to
set ansible_connection to 'ssh'. This patch changes it to
'local'.

Change-Id: I7494a937749a434f9bfb2e14c47b36aebdf89664
Closes-Bug: #1864649
(cherry picked from commit 5f0c07943eeb2c655f147be5e59b5128775ec34a)