Implement concurrency in baremetal cleaning
This change implements the concurrency functionality in the os_baremetal_clean_node
module. The module will now process all nodes provided concurently limiting threads
to the number of nodes or the option `concurrency` which has a default of 20. If the
number of nodes less that value of `concurrency` the thread pool will be limited
to the number of nodes.

Change-Id: I07f3c731c79cfe9bac8652f082e291832dc9b351
Signed-off-by: Kevin Carter <kecarter@redhat.com>