Add raid configuration playbook and options
This change adds the baremetal raid configuration option to the
os_baremetal_clean_node module. This was added here because
cleaning has a direct connection to raid configuration within the
API and CLI so it makes sense to couple these things within the
same module. The module will now handle the raid_config option
and apply the configuration to all node_uuid(s) provided prior
to running a cleaning event.

Story: 2007212
Task: 38449

Change-Id: I51b1977b53995c93eb661f9ffa6d7cbad033ab75
Signed-off-by: Kevin Carter <kecarter@redhat.com>