Only add host entries for role_networks
When network isolastion was added there was no have
composable networks and no jinja2 generated heat
templates. Because of this the "standard" networks
was always included, but the port template was
overridden to use the ctlplane's ip address for
disabled networks.

The result is that we have kept host entries for
networks that are irrelevant for some roles. This
was fixed in https://review.opendev.org/614457 but
re-intruduced in tripleo-ansibe.

This fixes the tripleo_hosts_entries role to only
write host entries for networks relavant to the
tripleo role.

Related-Bug: #1842001
Related-Bug: #1800811
Change-Id: I2c8f0eb33be26ebe6d4077b25de47c2b2ab1fca0