Creates container_config_data module
container_config_data is a new module that replaces some tasks that were
previously used to read and process JSON files on a host.

Doing it in a module will reduce the number of tasks, and make it a
little bit faster and easier to debug in big deployments.

Unit tests will come later.

Change-Id: I9844ed166ecf0718935ac1537b719000da816dbb