Auto-detect default overcloud-full image
When the undercloud glance is switched off the image files will be
copied to /var/lib/ironic/images. This change will choose what to set
for the default image based on whether there are image files in
/var/lib/ironic/images, falling back to a glance based
'overcloud-full'.

A missing .initrd file is used to imply that the .qcow2 file is a
whole-disk image.

This is just to determine the default image, and the image can always
be specified on a per-role basis in the baremetal provisioning yaml.

Change-Id: I3b8ba4b2a0d50077fcf4628720dff9310a200e2a
Blueprint: nova-less-deploy