Fail when there are no nodes to introspect
The mistral workflow would fail when no nodes were in a manageable
state to introspect. This change replicates that behaviour.

One of the causes of bug #1866204 is that there is a reliance on the
time it takes to to the introspect for the node segment mappings to
populate. When introspect finishes instantly, provide ends up timing
out because the port create fails.

Change-Id: I8a27a098b3cd006f73c33d95f2d93c65cf629203
Closes-Bug: #1866911
Related-Bug: #1866204