Remove mistral cli from the backup playbook
The undercloud backup playbook was using mistral to pull the mysql root
password from the available environment. While this works, its not needed
because the authoratative source is already in the deployment users home
folder. This change removes the use of the mistral cli calls, it will now
read the generated password file from the deployers home folder. The
playbook will read from the `tripleo-undercloud-passwords.yaml` file
first, if there are any issues with that file it will fallback to
`undercloud-passwords.conf`.

This removes our usage of mistral and speeds up the total playbook
runtime.

Story: 2007414
Task: 39060

Change-Id: I8bc3207129b60fca6e3205106bb967e414ce4984
Signed-off-by: Kevin Carter <kecarter@redhat.com>