Filter out wrapper commands from the ps output
This patch is filtering out the wrapper execution from the ps output
in the sync script. By doing this, it'll effectively detect when
the target process is not running and start it. Otherwise, there might
be cases where the process start is postponed until next iteration
of the sync script (1 minute) and it may be already too late.

This is causing tests to fail as the metadata service is not provisioned
in time for instances to fetch their SSH keys.

Change-Id: I530e257f343ffc551db9e984f9a27b20c397bfb1
Co-Authored-By: Jakub Libosvar <jlibosva@redhat.com>
Closes-Bug: #1868082
Signed-off-by: Daniel Alvarez <dalvarez@redhat.com>
(cherry-picked from 0df3ebd7592eac90f76650770f59c89fba4ebe39)