Fix the docs job
This change fixes the docs job to ensure we have access to development
versions of tripleo-common. This is being done because we have more
modules coming in which are using tripleo-common which may require
unreleased versions of the package when running tests. This change
adds tripleo-common as an upstream constraint which is only used in
the docs and molecule test scenarios.

Change-Id: Ib64307defe4d5a0e06ad862ab9f51d1da3d23312
Signed-off-by: Kevin Carter <kecarter@redhat.com>