Rework how we push the content in the file
Since we allow to pass lists, we can't just use |join('').

In addition, using the multi-line support in YAML via the "|" creates
some issues when we want to use |join('\n'): it gets escaped, meaning we
end with a one-line content with some \n entries.

The present patch ensures that:
- generated content is still pushed in a fact to not lose time
- static content (lists) is properly treated and outputs correctly

Change-Id: I7e520035f24503e59840b34e5b2d9e9966cd6ca4