Create a new playbook that will be used for deriving parameters
This change introduces a playbook that will be built on-top of to replace the
workflow `tripleo.derive_params.v1.derive_parameters`. At this time the playbook
is a reference implementation has no ability to compute specific fomulas used
to dervice HCI and NFV parameters, however, in follow commits the needed
functionality will be built out.

Story: 2007419
Task: 39120

Depends-On: I26ea1c01e17513d001e311d68fd239d3ba8c8a0d
Change-Id: I197278211152a589572e789f98ed68631ab20ee5
Signed-off-by: Kevin Carter <kecarter@redhat.com>