tripleo_container_manage: improve logging for failed containers
If a container fails to start after many retries, the default logging of
the async_status tasks isn't great and it's hard to figure out what
container failed to start.

In this patch, we introduce a new filter that will read the
async_results and build a list of containers which failed to start
(failed to True) or did not finish to start (finished to 0); the
async_status ignores errors, but we fail a bit later after building that
list.

Change-Id: I5a2270130bdf5b9d781f4d81ec25c6ccf12fdc07
(cherry picked from commit 8fdf2f276f505dbad80391c7e69b8c26fafbd303)