Add dynamic concurrency values
This change computes the concurrency values against the execution workload
allow the playbook to adjust the concurrency value to whatever the user has
defined when that value makes sense. If the concurrency value is greater
than the expected workload the value will be adjusted down as needed. While
this change is not 100% required it will ensure that the execution of our
playbooks is more programatically efficient.

Change-Id: I368187f142560108fe66a59f9acc2d9a6fac97d3
Signed-off-by: Kevin Carter <kecarter@redhat.com>