Normalize how we run validations
This change normalizes how we execute validations by ensuring all validations calls are
running via a shell command and sourcing the `stackrc` file; this is due to validations
not being completely compatible with the OSC clouds config. All validations will now
run when the `run_validations` is set to true. This ensures that the playbook will only
execute validations when instructed to.

Change-Id: Ie1b0c739dc08023a1bba7a98292a8c570b8a2a9c
Signed-off-by: Kevin Carter <kecarter@redhat.com>