Improve deployer UX by limiting failure output potential
This change will improve the deployer UX by ensuring that the log output does not show
failed tasks. While the potential for failed tasks should be rescued, and complete
successfully, the output confusing and could result in the operator taking unnecessary
debugging steps during a deployment. To ensure that the output is accurate and easy
to understand the playbook will now stat the required file and the needed blocks
will only execute when the appropriate conditions are met.

Change-Id: I0e69f44a6e06926a8987defa96c7ffac167ccdb5
Signed-off-by: Kevin Carter <kecarter@redhat.com>