Fix get_failed_containers() to correctly fetch container_data
To avoid this error:
The error was: KeyError: 'container_data'

We need to fetch container_data from async_result_item in the async
results; that's where the key is. Updating unit tests as well, and the
task which creates the facts so there is no confusion in the logs.

Change-Id: I2a5533335151c4b292e85aea310adfdc44ab1e02
(cherry picked from commit d6204498d7e9478a0ab36ef1b97acd2afeaf6d19)