Reduce overhead of stack show loop
A stack show of the overcloud is an expensive operation because
resolving outputs results in a nested stack traversal. This change
adds --no-resolve-outputs to the stack show loop which is waiting for
the stack to no longer be IN_PROGRESS.

On my undercloud this reduces stack show time from 14 seconds to 3
seconds. Since polling load is lower now, the polling period is
changed to a shorter 10 seconds.

Change-Id: I34dcf67b36d802f9ccfb65741012e0ed32a15077