tripleo_container_rm: fix check mode
If tripleo_container_rm is run in check mode, the task which detect what
container cli is used will be skipped and the next task relying on the
register will fail to run.

Let's force the first tasks to be run so we know which container cli is
used; and keep running the rest of the playbooks in check mode.

Change-Id: I4c5f92b7f308d655c5c94c62c694ad1cea08d625