Add connection check before configuring tripleo-admin
This change adds a connection check to the "Run create admin" playbook
which will ensure the target host is up and ready to recieve connection
prior to attempting to execute the rest of the playbook. This is being
done because there is a chance servers may not be up and available by
the time this playbook is executed and we dont want to cause a bad
user experience by failing early, forcing the deployer to rerun a
deployment for no reason.

Closes-Bug: #1873246
Change-Id: I2d2e3bdcaf0a31d8daf90656533ac0f55b70fa26
Signed-off-by: Kevin Carter <kecarter@redhat.com>