tripleo_container_manage: add safeguard against wrong healthcheck config
If a container config has by mistake a healthcheck but no systemd
restart policy, we don't want to manage the healthcheck because it
requires its service to be created.

To prevent that situation, we'll create the healthchecks only if they
are already part of the systemd services list that was created earlier.
For that, we're using the intersect() filter which allows to
get the intersection of 2 lists (systemd services and healthchecks to
create).

Adding molecule coverage to test this scenario.

Closes-Bug: #1873249
Change-Id: Id5cc784bae597def0648f07d28b6463b387d2212