tripleo_container_manage: manage systemd services per step vs start_order
Instead of managing systemd services per start_order per step, we could
manage them per step.

The start_order was created to be able to create containers which runs
some command or do exec; but not really for the ones who are services
and managed by systemd.

Doing it per step will reduce the number of tasks and therefore the
deployment time.

Note: it adds dict_to_list() filter; which converts a dict of dicts to a
list of dicts. Ansible allows to do it via dict2items | list but in this
particular case we don't want key/value when later treating the data to
figure out if systemd is needed.

Change-Id: Ia38f2ec753dc3c21bcf91f057fe7ff8020d214e6
(cherry picked from commit ca80ec5c17623935a0479bbf198484fc873e977c)