tripleo-redhat-enforce new role to enforce RHOSP constraints.
OSP, the downstream version of tripleo have to enforce some policies
for rhel version and subscribed channel.

This module hosts those requirements, so that we prevent update to
wrong rhel release or subscription to wrong channels.

Currently it only implements a basic check to the subscribed rhel
version.

This check has some fail-safe logic to avoid crashing the update on
temporary network issue when running subscription-manager.

We are avoiding the validation framework as this can be easily
disabled and we want this enforcement to be mandatory as this could
lead user to unsupported combination of OSP/RHEL.

For upstream that change is transparent as the tasks are skipped if
the ansible_distribution is not Red Hat.

For Red Hat, there will be a mechanism in THT to avoid the check
altogether, for instance for CI purposes.

For this first check (RHEL/OSP version), downstream patches will add
the required values in vars/redhat.yml.

Note about the backport, there is a role naming change relative to
master, we use "-" instead of "_" in train for the role directory.
The role directory has to be renamed. Documentation is pointing
to molecule/playbook.yml and not molecule/converge.yml.

Change-Id: I2d1ac92ee6ee8407fb156a2718f94ad3e9220bbe
(cherry picked from commit e65996b878a9c758d5b41e996776c68cfe686d29)