Add tripleo callback
The default calback has the worst UX messaging for async tasks ever.
We're creating a tripleo callback that can be configured to replace the
standard 'FAILED - RETRYING: %s (%d retries left).' with the more
userfriendly 'WAITING FOR COMPLETION: %s (%d retries left).' when it's
an async task and a 'RETRY: %s (%d retries left).' when it's a failure
retry.

The tripleo callback inherits everything else from the ansible default
callback that is currently used and overrides the individual functions
as necessary.

See also https://github.com/ansible/ansible/issues/63926

Change-Id: I3582c5ed80f755e2206782823457338d1fb42213